\name{Description}
\alias{Description}
\alias{Description.APP_DESCRIPTION}
\alias{Description.APP_NAME}
\alias{Description.DEPT_NAME}
\alias{Description.DSGROUP_NAME}
\alias{Description.INDIVIDUAL_NAME}
\alias{Description.SERVICE_GROUP_ID}
\alias{Description.SERVICE_GROUP_NAME}
\alias{Description.SERVICE_NAME}
\alias{Description.SERVICE_TYPE_NAME}
\alias{Description.USERNAME}
\docType{data}
\title{
Description properties
}
\description{
Defined Service description properties. They can be used for accounting and discrimination.
}
\value{
\item{Description.APP_DESCRIPTION }{appDescription, the application description for accounting}
\item{Description.APP_NAME }{appName, the application name for accounting}
\item{Description.DEPT_NAME }{deptName, the department name for accounting}
\item{Description.DSGROUP_NAME }{groupName, the group name for accounting}
\item{Description.INDIVIDUAL_NAME }{individualName, the individual name for accounting}
\item{Description.SERVICE_GROUP_ID }{The ID of the Service Group, which is automatically set by the \code{ServiceGroup} object.}
\item{Description.SERVICE_GROUP_NAME }{The name of the Service Group.}
\item{Description.SERVICE_NAME }{serviceName, the Service name, which is the Service ID by default. This is displayed as the Session Name on the \code{Service Session Admin} page.}
\item{Description.SERVICE_TYPE_NAME }{serviceTypeName, the name of the Service type.  This is the registered name used to create the Service.}
\item{Description.USERNAME }{The client's username. Automatically set.}
}
\seealso{
}
\keyword{datasets}
