\name{rdriver-package}
\alias{rdriver-package}
\alias{rdriver}
\docType{package}
\title{rdriver-package
R driver package
}
\description{
Allows R application to distribute workloads to Engines running in the Grid.
}
\details{
\tabular{ll}{
Package: \tab rdriver\cr
Type: \tab Package\cr
Version: \tab 6.2.0.159175\cr
Date: \tab 2017-11-21 13:31\cr
License: \tab GridServer license.\cr
}
}
\author{
TIBCO Software Inc.
DSPG
}
\keyword{ package }
\seealso{
\code{\link[rlogger:rlogger-package]{rlogger}}
}
\examples{
}
